[Decomp][Pet] Cohesion split: extract persistence + spells/talents to Pet*.cpp
Mirror of mangosthree #265. Pure cohesion split of Pet.cpp with no
behaviour change:

- PetDatabase.cpp  - pet persistence: load from / save to / delete from
  the characters database.
- PetSpells.cpp    - pet spell book, spell/aura/cooldown persistence,
  talents (reset/cost/level) and autocast toggling.

Pet.h is unchanged. CMake file(GLOB Object/*.cpp) auto-compiles the new
translation units.
